+++ internal/ui/frame.go
@@ -2,8 +2,6 @@ package ui
 
 import "charm.land/lipgloss/v2"
 
-// frame はパネルの枠線。通常・選択中・フォーカス中の 3 状態を
-// 罫線の太さと色で区別する。幅は 3 状態とも 1 セルで変わらない。
 type frame struct {
 	topLeft     string
 	topRight    string
@@ -63,8 +61,6 @@ func (box frame) render(value string) string {
 	return box.style.Render(value)
 }
 
-// 枠の色は applyTheme がプリセットから差し替える。罫線の太さは状態ごとに
-// 固定で、設定で変わるのは色だけ。
 func (model *Model) frameFor(target panel) frame {
 	if model.panel != target {
 		return plainFrame
